+++ src/core/generators/axios.ts
@@ -74,18 +74,21 @@ const generateImports = ({
 export const generateAxiosTitle = (title: string) => {
   const sanTitle = sanitize(title);
   return {
-    definition: `I${pascal(sanTitle)}Api`,
-    implementation: `get${pascal(sanTitle)}Api`,
-    implementationMock: `get${pascal(sanTitle)}ApiMock`,
+    definition: `${pascal(sanTitle)}`,
+    implementation: `get${pascal(sanTitle)}`,
+    implementationMock: `get${pascal(sanTitle)}Mock`,
   };
 };
 
-export const generateAxiosHeader = (title: string) => {
-  const axiosTitle = generateAxiosTitle(title);
+export const generateAxiosHeader = (titles: {
+  definition: string;
+  implementation: string;
+  implementationMock: string;
+}) => {
   return {
-    definition: `export interface ${axiosTitle.definition} {`,
-    implementation: `export const ${axiosTitle.implementation} = (axios: AxiosInstance): ${axiosTitle.definition} => ({\n`,
-    implementationMock: `export const ${axiosTitle.implementationMock} = (): ${axiosTitle.definition} => ({\n`,
+    definition: `export interface ${titles.definition} {`,
+    implementation: `export const ${titles.implementation} = (axios: AxiosInstance): ${titles.definition} => ({\n`,
+    implementationMock: `export const ${titles.implementationMock} = (): ${titles.definition} => ({\n`,
   };
 };